ScheduleTreeExtension: hide copy constructor
ScheduleTree nodes are not supposed to be manipulated by-value, only via unique
pointers.  Make the copy constructor of ScheduleTreeExtension private,
similarly to its main constructor, and provide a static member function to
return copies wrapped in a unique pointer.  Note that unlike the conventional
copy constructor, this static function takes a pointer to the tree rather than
a reference.  This keeps the API consistent as all other functions manipulate
with pointers to the tree nodes rather than references.